+++ src/GRA.Domain.Repository/IUserRepository.cs
@@ -48,7 +48,7 @@ namespace GRA.Domain.Repository
         Task<ICollection<User>> GetHouseholdUsersWithAvailablePrizeAsync(
             int headId, int? drawingId, int? triggerId);
 
-        Task<IEnumerable<int>> GetNewsSubscribedUserIdsAsync(int siteId);
+        Task<IEnumerable<int>> GetNewsSubscribedUserIdsAsync(int siteId, bool excludeCannotBeEmailed);
 
         Task<IDictionary<User, int>> GetStaffRegisteredParticipantsAsync(ReportCriterion criterion);
 
@@ -90,6 +90,8 @@ namespace GRA.Domain.Repository
             int take);
 
         Task<int> ReassignBranchAsync(int oldBranchId, int newBranchId);
+
+        Task SetCannotBeEmailedAsync(int currentUserId, int userId, bool cannotBeEmailed);
 
         Task SetUserPasswordAsync(int currentUserId, int userId, string password);