+++ src/GRA/SegmentNames.cs
@@ -2,6 +2,7 @@
 {
     public static class SegmentNames
     {
+        public const string ProgramButtonText = "Program Button Text";
         public const string VendorCodeDonation = "Vendor Code Donation";
         public const string VendorCodeEmailAward = "Vendor Code Email Award";
     }